### 2018-09-14 #944 @arfon
@leouieda - this PR https://github.com/fatiando/verde/pull/135 will fix the paper compilation issue.
