### 2018-09-14 91cf7c8
LaTeX fix for JOSS paper (#135)
