### 2018-09-14 #944 @leouieda
@arfon thanks for the fix! It's merged now.
