### 2018-09-17 #944 @lheagy
:wave: Hi @sgrieve, @fil, @Chilipp, @WalBytes, would one of you be willing to review this submission for JOSS?
