### 2018-09-17 #944 @Fil
> 👋 would one of you be willing to review this submission for JOSS?

yes, thanks!
