### 2018-09-17 #957 @whedon
[REVIEW]: Verde: Processing and gridding spatial data using Green's functions
**Submitting author:** @leouieda (<a href="http://orcid.org/0000-0001-6123-9515">Uieda, L</a>)
**Repository:** <a href="https://github.com/fatiando/verde" target ="_blank">https://github.com/fatiando/verde</a>
**Version:** v1.0.0
