### 2018-09-17 #957 @lheagy
Many thanks @fourndo, @Fil for being willing to review 🎉. Above, there is a checklist for each of you to work through to guide the review. Please let me know if you have any questions or if I can provide any clarification!
