### 2018-09-17 #957 @leouieda
Hi @fourndo and @Fil, thanks for taking the time to review! Just to let you know that conda installation isn't working right now because of conda-forge/staged-recipes#6659. I'll let you know once that is resolved and we have a working conda package.
