### 2018-09-19 bc9e28d
fixed a typo
just fixed a typo in the documentation
