### 2018-09-19 c2fb3b0
Force matplotlib<3.0.0 because of Cartopy (#137)
Cartopy is breaking under matplotlib 3.0.0. Require <3.0.0 until they
sort it out.
