+++ environment.yml
@@ -17,7 +17,7 @@ dependencies:
     # Development requirements
     - dask
     - pyproj
-    - matplotlib
+    - matplotlib<3.0.0
     - cartopy
     - jupyter
     - pytest
