### 2018-09-21 #957 @leouieda
The error is resolved and packages are now build and available through conda-forge.
