### 2018-09-21 48b34e1
Update UH affiliation on the JOSS paper
The department changed names from "Geology and Geophysics"
to "Earth Sciences".
