### 2018-09-21 #957 @Fil
I have installed and tested it — no problem! I just would have preferred to start a jupyter notebook rather than tutorials/*py, but that's a personal preference.
