### 2018-09-21 #957 @leouieda
The tutorials are converted to notebooks by sphinx-gallery. There's a
button to download them at the bottom of each page, like
https://www.fatiando.org/verde/latest/tutorials/overview.html#sphx-glr-download-tutorials-overview-py

Notebooks are so hard to maintain in version control. I really like this
option of editing a .py but being able to get notebooks in the end :)
