### 2018-09-27 #957 @fourndo
@lheagy Where is the ```paper.md``` supposed to be?
