### 2018-09-27 #957 @lheagy
hi @fourndo, the [paper.md](https://github.com/fatiando/verde/blob/master/paper/paper.md) is in the repo, but it is nicer to look at the [pdf version](https://github.com/openjournals/joss-papers/blob/joss.00957/joss.00957/10.21105.joss.00957.pdf) that will be used
