### 2018-09-27 #957 @lheagy
Many thanks @fourndo and @Fil for your speedy reviews 🎉

@leouieda: congrats on a nice piece of work 😄 Can you please archive the package on [Zenodo](https://zenodo.org) or similar and post the doi here?
