### 2018-10-05 #957 @leouieda
@lheagy I already had a Zenodo archive for v1.0.0. Since there weren't changes requested, can we use this one? https://doi.org/10.5281/zenodo.1415280  The only new thing in the repo is the paper.
