### 2018-10-06 #957 @lheagy
Thanks @leouieda, yes, that works. Would you mind doing a new release with the final JOSS paper to capture the current state of the repo?
