+++ verde/tests/test_datasets.py
@@ -1,12 +1,7 @@
-# pylint: disable=wrong-import-position
 """
 Test data fetching routines.
 """
-# Import matplotlib and set the backend before anything else to make sure no windows are
-# created and there are no problem with TravisCI running in headless mode.
 import matplotlib
-
-matplotlib.use("agg")
 import matplotlib.pyplot as plt
 import cartopy.crs as ccrs
 
@@ -22,6 +17,11 @@ from ..datasets.sample_data import (
     fetch_texas_wind,
     setup_texas_wind_map,
 )
+
+
+# Set the backend before anything else to make sure no windows are created and there are
+# no problem with TravisCI running in headless mode.
+matplotlib.use("agg")
 
 
 def test_fetch_texas_wind():
