+++ environment.yml
@@ -26,6 +26,7 @@ dependencies:
     - coverage
     - black
     - pylint
+    - flake8
     - sphinx
     - sphinx_rtd_theme
     - sphinx-gallery
