### 2018-10-10 f269bb7
Don't install dependencies when only checking
