+++ .travis.yml
@@ -43,8 +43,8 @@ matrix:
               - CHECK=true
               # Don't need to install everything just to run the style checks
               - CONDA_INSTALL_EXTRA="black flake8 pylint"
-              - CONDA_REQUIREMENTS=
-              - CONDA_REQUIREMENTS_DEV=
+              - CONDA_REQUIREMENTS=""
+              - CONDA_REQUIREMENTS_DEV=""
         # Main build to deploy to PyPI and Github pages
         - os: linux
           env:
