### 2018-10-10 2baed5a
Specify the matplotlib backend on the command line
