+++ verde/tests/test_datasets.py
@@ -1,7 +1,6 @@
 """
 Test data fetching routines.
 """
-import matplotlib
 import matplotlib.pyplot as plt
 import cartopy.crs as ccrs
 
@@ -17,11 +16,6 @@ from ..datasets.sample_data import (
     fetch_texas_wind,
     setup_texas_wind_map,
 )
-
-
-# Set the backend before anything else to make sure no windows are created and there are
-# no problem with TravisCI running in headless mode.
-matplotlib.use("agg")
 
 
 def test_fetch_texas_wind():
