### 2018-10-10 ce0a224
Remove default shape for BaseGridder.grid
There is no reason to have one and it wasn't even implemented in
`grid_coordinates`. This only affected a single example of CheckerBoard
that didn't set `shape` or `spacing`.
