+++ setup.cfg
@@ -8,3 +8,6 @@ tag_prefix = ''
 [flake8]
 ignore = E203, E266, E501, W503, F401, E741
 max-line-length = 88
+max-complexity = 10
+exclude =
+    verde/_version.py
