+++ .travis.yml
@@ -27,41 +27,47 @@ env:
         - CONDA_REQUIREMENTS=requirements.txt
         - CONDA_REQUIREMENTS_DEV=requirements-dev.txt
         # These variables control which actions are performed in a build
-        - COVERAGE=false
+        - TEST=false
+        - CHECK=false
         - BUILD_DOCS=false
-        - CHECK_FORMAT=false
         - DEPLOY_DOCS=false
         - DEPLOY_PYPI=false
 
 # Specify the build configurations. Be sure to only deploy from a single build.
 matrix:
     include:
+        # Code style and quality checks
         - os: linux
           env:
               - PYTHON=3.6
-              - COVERAGE=true
+              - CHECK=true
+              # Don't need to install everything just to run the style checks
+              - CONDA_INSTALL_EXTRA="black flake8 pylint"
+              - CONDA_REQUIREMENTS=""
+              - CONDA_REQUIREMENTS_DEV=""
+        # Main build to deploy to PyPI and Github pages
+        - os: linux
+          env:
+              - PYTHON=3.6
+              - TEST=true
               - BUILD_DOCS=true
               - DEPLOY_DOCS=true
               - DEPLOY_PYPI=true
-              - CHECK_FORMAT=true
-              # Black is Python 3.6 only so it can't be in the requirements file.
-              - CONDA_INSTALL_EXTRA="black"
         - os: linux
           env:
               - PYTHON=3.5
+              - TEST=true
               - BUILD_DOCS=true
+        # Check tests pass when using optional dependencies.
         - os: linux
           env:
               - PYTHON=3.6
-              - COVERAGE=true
+              - TEST=true
               - CONDA_INSTALL_EXTRA="pykdtree numba"
         - os: osx
           env:
               - PYTHON=3.6
-              - COVERAGE=true
-        - os: osx
-          env:
-              - PYTHON=3.5
+              - TEST=true
               - BUILD_DOCS=true
 
 # Setup the build environment
@@ -86,14 +92,13 @@ install:
 # Run the actual tests and checks
 script:
     # Check code for style and quality
-    - if [ "$CHECK_FORMAT" == "true" ]; then
+    - if [ "$CHECK" == "true" ]; then
           make check;
+          make lint;
       fi
     # Run the test suite
-    - if [ "$COVERAGE" == "true" ]; then
+    - if [ "$TEST" == "true" ]; then
           pip install codecov;
-          make coverage;
-      else
           make test;
       fi
     # Build the documentation
@@ -104,7 +109,7 @@ script:
 # Things to do if the build is successful
 after_success:
     # Upload coverage information
-    - if [ "$COVERAGE" == "true" ]; then
+    - if [ "$TEST" == "true" ]; then
         coverage xml;
         echo "Uploading coverage to Codecov";
         codecov -e PYTHON;
