### 2018-10-10 0f16e8b
Remove default shape for BaseGridder.grid (#140)
There is no reason to have one and it wasn't even implemented in
`grid_coordinates`. This only affected a single example of CheckerBoard
that didn't set `shape` or `spacing`.
