### 2018-10-10 #957 @leouieda
@lheagy OK, I have made all the updates I needed. If the citation in the preview PDF going to be the final one? If so, I'd like to include it in the CITATION file before making the release (this way I don't need a new version just for updating that).
