### 2018-10-11 b36731c
Changes for 1.0.1 release and JOSS citation
Update citation to reflect the JOSS paper (accepted). Update links on
README and include a changelog.
