### 2018-10-11 9bb9538
Changes for 1.0.1 release and JOSS citation (#142)
Update citation to reflect the JOSS paper (accepted). Update links on
README and include a changelog.
