+++ CITATION.rst
@@ -4,18 +4,24 @@ Citing Verde
 This is research software **made by scientists**. Citations help us justify the effort
 that goes into building and maintaining this project.
 
-If you used Verde in your research, please consider citing us:
+If you used Verde in your research, please consider citing our paper:
 
-    Uieda, L., and Hoese, D. (2018). Verde v1.0.0: Processing and gridding spatial data using Green's
-    functions. Zenodo. https://doi.org/10.5281/zenodo.1415281
+    Uieda, L. (2018). Verde: Processing and gridding spatial data using Green's
+    functions. Journal of Open Source Software, 3(29), 957. doi:10.21105/joss.00957
+
+This is an open-access publication and can be freely downloaded from
+https://doi.org/10.21105/joss.00957
 
 Here is a Bibtex entry to make things easier if you're using Latex::
 
-    @misc{verde,
-      author       = {Leonardo Uieda and David Hoese},
-      title        = {Verde v1.0.0: Processing and gridding spatial data using Green's functions},
-      year         = 2018,
-      publisher    = {Zenodo},
-      doi          = {10.5281/zenodo.1415281},
-      url          = {https://doi.org/10.5281/zenodo.1415281}
+    @article{uieda2018,
+      title = {{Verde}: {Processing} and gridding spatial data using {Green's} functions},
+      author = {Uieda, Leonardo},
+      year = {2018},
+      journal = {Journal of Open Source Software},
+      volume = {3},
+      number = {29},
+      pages = {957}
+      issn = {2475-9066},
+      doi = {10.21105/joss.00957},
     }
