### 2018-10-11 #957 @leouieda
I had a look at other papers and they all used the same DOI and citation as the preview so I went ahead and made a new release (v1.0.1).

This is the Zenodo archive of it: https://doi.org/10.5281/zenodo.1421979

@lheagy we're good to go on my part. Thanks for editing! :1st_place_medal:
