### 2018-10-11 bfae87b
Deploy v1.0.1 from TravisCI
