### 2018-10-11 1dcbc2a
Set default n_splits=5 in scikit-learn KFold (#143)
scikit-learn will change the default from 3 to 5 in an upcoming release.
Set the value to 5 now to prepare and make sure that future changes
don't affect us.
