+++ verde/base.py
@@ -236,50 +236,38 @@ class BaseGridder(BaseEstimator):
         region=None,
         shape=None,
         spacing=None,
-        adjust="spacing",
         dims=None,
         data_names=None,
         projection=None,
+        **kwargs,
     ):
         """
         Interpolate the data onto a regular grid.
 
-        The grid can be specified by either the number of points in each
-        dimension (the *shape*) or by the grid node spacing.
+        The grid can be specified by either the number of points in each dimension (the
+        *shape*) or by the grid node spacing. See :func:`verde.grid_coordinates` for
+        details. Other arguments for :func:`verde.grid_coordinates` can be passed as
+        extra keyword arguments (``kwargs``) to this method.
 
-        If the given region is not divisible by the desired spacing, either the
-        region or the spacing will have to be adjusted. By default, the spacing
-        will be rounded to the nearest multiple. Optionally, the East and North
-        boundaries of the region can be adjusted to fit the exact spacing
-        given. See :func:`verde.grid_coordinates` for more details.
+        If the interpolator collected the input data region, then it will be used if
+        ``region=None``. Otherwise, you must specify the grid region.
 
-        If the interpolator collected the input data region, then it will be
-        used if ``region=None``. Otherwise, you must specify the grid region.
-
-        Use the *dims* and *data_names* arguments to set custom names for the
-        dimensions and the data field(s) in the output :class:`xarray.Dataset`.
-        Default names are provided.
+        Use the *dims* and *data_names* arguments to set custom names for the dimensions
+        and the data field(s) in the output :class:`xarray.Dataset`. Default names will
+        be provided if none are given.
 
         Parameters
         ----------
         region : list = [W, E, S, N]
-            The boundaries of a given region in Cartesian or geographic
-            coordinates.
+            The boundaries of a given region in Cartesian or geographic coordinates.
         shape : tuple = (n_north, n_east) or None
             The number of points in the South-North and West-East directions,
-            respectively. If *None* and *spacing* is not given, defaults to
-            ``(101, 101)``.
-        spacing : tuple = (s_north, s_east) or None
-            The grid spacing in the South-North and West-East directions,
             respectively.
-        adjust : {'spacing', 'region'}
-            Whether to adjust the spacing or the region if required. Ignored if
-            *shape* is given instead of *spacing*. Defaults to adjusting the
-            spacing.
+        spacing : tuple = (s_north, s_east) or None
+            The grid spacing in the South-North and West-East directions, respectively.
         dims : list or None
             The names of the northing and easting data dimensions, respectively, in the
-            output grid. Defaults to ``['northing', 'easting']`` for Cartesian grids and
-            ``['latitude', 'longitude']`` for geographic grids. **NOTE: This is an
+            output grid. Defaults to ``['northing', 'easting']``. **NOTE: This is an
             exception to the "easting" then "northing" pattern but is required for
             compatibility with xarray.**
         data_names : list of None
@@ -300,19 +288,17 @@ class BaseGridder(BaseEstimator):
         Returns
         -------
         grid : xarray.Dataset
-            The interpolated grid. Metadata about the interpolator is written
-            to the ``attrs`` attribute.
+            The interpolated grid. Metadata about the interpolator is written to the
+            ``attrs`` attribute.
 
         See also
         --------
         verde.grid_coordinates : Generate the coordinate values for the grid.
 
         """
-        dims = get_dims(self, dims)
+        dims = get_dims(dims)
         region = get_instance_region(self, region)
-        coordinates = grid_coordinates(
[…]
