### 2018-10-11 cb805e6
Remove trailing commas and old docs
