### 2018-10-11 217ba86
Allow gridders to pass kwargs to coord generators (#144)
Gridder methods `grid`, `scatter`, and `profile` now take in `**kwargs`
and pass them along to the coordinate generation functions. This will
allow extra flexibility for passing in arguments, like a grid height or
configuration, without needing to keep the methods in sync with the
functions.

Removed the `coordinate_system` mechanics from `profile` because it
wasn't implemented and is probably gonna take a while. Will introduce it
back when we finally have a geographic coordinate gridder. This
simplifies the coordinate naming schemes and
[…]
